django/rest_framework_bpmappers/myproject/requirements.txt
Django==1.5.4
bpmappers==0.7
djangorestframework==2.3.8
